[Refactor] 솔님꺼 풀 받기 전 커밋
