+++ src/main/java/project/moseup/controller/MemberController.java
@@ -10,6 +10,8 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import project.moseup.domain.Member;
 import project.moseup.domain.MemberGender;
+import project.moseup.dto.JoinForm;
+import project.moseup.dto.LoginForm;
 import project.moseup.service.MemberService;
 
 import javax.validation.Valid;
