+++ src/main/java/project/moseup/dto/TeamAskForm.java
@@ -1,4 +1,4 @@
-package project.moseup.controller;
+package project.moseup.dto;
 
 import lombok.Getter;
 import lombok.Setter;
