--- src/main/resources/templates/fragments/navbar.html
@@ -1,46 +1,0 @@
-<html lang="ko" xmlns:th="http://www.thymeleaf.org">
-<div th:fragment="navbar" id="navbar">
-	<nav class="p-3 mb-3 border-bottom">
-		<div class="nav-container">
-			<div class="d-flex flex-wrap align-items-center justify-content-center justify-content-lg-start nav-width">
-				<!-- 로고 이미지 -->
-				<a href="/" class="d-flex align-items-center mb-2 mb-lg-0 text-dark text-decoration-none nav-left">
-					<img class="bi me-2" width="45" height="35" alt="moseup" src="/images/logo.png">
-				</a>
-				
-				<!-- 메뉴 리스트 -->
-				<ul class="nav col-12 col-lg-auto me-lg-auto mb-2 justify-content-center mb-md-0">
-					<li><a href="#" class="nav-link px-2 link-secondary">menu1</a></li>
-					<li><a href="#" class="nav-link px-2 link-dark">menu2</a></li>
-					<li><a href="#" class="nav-link px-2 link-dark">menu3</a></li>
-					<li><a href="#" class="nav-link px-2 link-dark">menu4</a></li>
-				</ul>
-	
-				<!-- 검색 폼 -->
-				<form class="col-12 col-lg-auto mb-3 mb-lg-0 me-lg-3" role="search">
-					<input type="search" class="form-control" placeholder="Search..."
-						aria-label="Search">
-				</form>
-				
-				<div class="dropdown text-end">
-					<!-- 비로그인(로그인/회원가입) -->
-					
-					<!-- 로그인(마이페이지) -->
-					<a href="#" class="d-block link-dark text-decoration-none dropdown-toggle nav-right"
-						id="dropdownUser1" data-bs-toggle="dropdown" aria-expanded="false">
-						<img src="/images/profile.png" alt="..." width="40" height="40" class="rounded-circle">
-					</a>
-					<ul class="dropdown-menu text-small" aria-labelledby="dropdownUser1">
-						<li><a class="dropdown-item" href="#">New team...</a></li>
-						<li><a class="dropdown-item" href="#">Settings</a></li>
-						<li><a class="dropdown-item" href="#">myPage</a></li>
-						<li><hr class="dropdown-divider"></li>
-						<li><a class="dropdown-item" href="#">Sign out</a></li>
-					</ul>
-				</div>
-			</div>
-		</div>
-	</nav>
-</div>
-</html>
-
