+++ src/main/resources/templates/layouts/fragments/adminCommon.html
